# cmd/hypercloud-server/app/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Hypercloud Server API.",
    "contact": {
      "name": "Axel Etcheverry",
      "email": "axel@etcheverry.biz"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "host": "localhost",
  "basePath": "/v1",
  "paths": {
...
    "/v1/services": {
      "get": {
        "description": "Get the services list",
        "tags": [
          "Service"
        ],
        "operationId": "getServicesHandler",
        "parameters": [
          {
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
            "type": "string",
            "name": "StackID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Service"
          }
        }
      },
      "post": {
        "description": "Create service\nRequest: ServiceCreateRequest",
        "tags": [
          "Service"
        ],
        "operationId": "postServiceHandler",
        "responses": {
          "201": {
            "$ref": "#/responses/Service"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Service": {
      "type": "object",
      "title": "Service represents a Service.",
      "required": [
        "name"
      ],
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "hosts": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Hosts"
        },
        "id": {
          "type": "string",
          "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
          "x-go-name": "ID"
        },
        "name": {
          "description": "Service name.",
          "type": "string",
          "minLength": 3,
          "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
          "x-go-name": "Name"
        },
        "stack_id": {
          "type": "string",
          "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
          "x-go-name": "StackID"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedAt"
        }
      },
      "x-go-package": "github.com/hyperscale/hypercloud/database/entity"
    },
...
  },
  "responses": {
    "Service": {
      "schema": {
        "$ref": "#/definitions/Service"
      }
    },
...
  }
}